Minor refactor, refactor temp view messagehandler
