+++ resources/templates/course-details.html
@@ -147,8 +147,7 @@
                     }
                 }
                 if (ids.length > 0) {
-                    vscode.postMessage({ type: type, ids, id: {{ courseId }}
-            });
+                    vscode.postMessage({ type: type, ids, id: {{ courseId }} });
                 }
             }
