+++ resources/templates/downloading-exercises.html
@@ -46,7 +46,7 @@
                                 aria-valuemin="0"
                                 aria-valuemax="{{total}}"
                             >
-                                {{successful_pct}} %
+                                {{successfulPct}} %
                             </div>
                             <div
                                 class="progress-bar bg-danger"
@@ -56,7 +56,7 @@
                                 aria-valuemin="0"
                                 aria-valuemax="{{total}}"
                             >
-                                {{failed_pct}} %
+                                {{failedPct}} %
                             </div>
                         </div>
                     </div>
